+++ svm/src/transaction_processing_callback.rs
@@ -1,17 +1,10 @@
-use {
-    solana_program_runtime::loaded_programs::ProgramCacheMatchCriteria,
-    solana_sdk::{account::AccountSharedData, pubkey::Pubkey},
-};
+use solana_sdk::{account::AccountSharedData, pubkey::Pubkey};
 
 /// Runtime callbacks for transaction processing.
 pub trait TransactionProcessingCallback {
     fn account_matches_owners(&self, account: &Pubkey, owners: &[Pubkey]) -> Option<usize>;
 
     fn get_account_shared_data(&self, pubkey: &Pubkey) -> Option<AccountSharedData>;
-
-    fn get_program_match_criteria(&self, _program: &Pubkey) -> ProgramCacheMatchCriteria {
-        ProgramCacheMatchCriteria::NoCriteria
-    }
 
     fn add_builtin_account(&self, _name: &str, _program_id: &Pubkey) {}
 }
